% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgteam.R
\name{avgteam}
\alias{avgteam}
\title{Find averages of stats}
\usage{
avgteam(teammatrix, w = FALSE)
}
\arguments{
\item{teammatrix}{A matrix of player stats created by buildteam()}

\item{w}{TRUE/FALSE for weighting stats by games played}
}
\value{
Table of average player stats
}
\description{
Find averages of stats
}
\examples{
team1 <- buildteam(c("Cr0", "Gladiator"))
avgteam(team1)
}
