% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestplayer.R
\name{bestplayer}
\alias{bestplayer}
\title{Find best player}
\usage{
bestplayer(team_matrix)
}
\arguments{
\item{teammatrix}{A matrix of player stats created by buildteam()}
}
\value{
Stats of the best player
}
\description{
Find best player
}
\examples{
team1 <- buildteam(c("Cr0", "Gladiator"))
bestplayer(team1)
}
