% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestteam.R
\name{bestteam}
\alias{bestteam}
\title{Find best team}
\usage{
bestteam(...)
}
\arguments{
\item{...}{A number of team matrices created by buildteam()}
}
\value{
Average stats of the best team
}
\description{
Find best team
}
\examples{
team1 <- buildteam(c("Cr0", "Gladiator"))
team2 <- buildteam(c("Cr0", "JellyG"))
bestplayer(team1, team2)
}
