% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildteam.R
\name{buildteam}
\alias{buildteam}
\title{Create table of player stats}
\usage{
buildteam(playerlist)
}
\arguments{
\item{playerlist}{A list of valid player names}
}
\value{
Table of player stats
}
\description{
Create table of player stats
}
\examples{
buildteam(c("Cr0", "Gladiator"))
}
