% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickteam.R
\name{quickteam}
\alias{quickteam}
\title{Create table of player stats but its easier to enter names}
\usage{
quickteam(...)
}
\arguments{
\item{...}{A list of valid player names}
}
\value{
Table of player stats
}
\description{
Create table of player stats but its easier to enter names
}
\examples{
quickteam(Cr0, Gladiator)
}
