% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplayers.R
\name{readplayers}
\alias{readplayers}
\title{Turn text file into list}
\usage{
readplayers(path)
}
\arguments{
\item{path}{Path to text file}
}
\value{
List of player names
}
\description{
Turn text file into list
}
\examples{
team1 <- readplayers("./team.txt")
}
